Updated trigger-warnings.php to read warnings.json
This route configures trigger-warnings.php to read the warnings from a
file. The intention behind storing the warnings within a JSON file is as
a stepping stone towards storing this data within the Wordpress
database, but to make the initial task of editing and updating warnings
as easy as possible in the short to medium term.